+++ src/components/little-palette/LittlePalette.tsx
