+++ src/components/palette/Palette.tsx
@@ -3,7 +3,7 @@ import {
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@components/ui/tooltip";
+} from "@/components/shared/ui/tooltip";
 import { clipboard } from "@utils/clipboard";
 import { memo, useState } from "react";
 import { toast } from "sonner";
