+++ src/components/shared/items/SliderComponent.tsx
