+++ src/components/shared/ui/alert-dialog.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
 
 import { cn } from "@lib/utils";
-import { buttonVariants } from "@components/ui/button";
+import { buttonVariants } from "@/components/shared/ui/button";
 
 const AlertDialog = AlertDialogPrimitive.Root;
