+++ src/components/shared/ui/slider.tsx
